requirements.txt
aiohttp==3.7.4.post0
argcomplete==2.0.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
discord.py==1.7.3
distlib==0.3.4
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
numpy==1.23.0
packaging==21.3
pandas==1.4.3
pipenv==2022.3.28
pipx==1.0.0
platformdirs==2.5.1
pluggy==1.0.0
praw==7.5.0
prawcore==2.3.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
requests==2.27.1
six==1.16.0
tomli==2.0.1
traitlets==5.2.2.post1
typer==0.4.1
typing_extensions==4.2.0
update-checker==0.18.0
urllib3==1.26.8
userpath==1.8.0
virtualenv==20.14.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
websocket-client==1.3.1
yarl==1.7.2
